[Fix] rename componets
